% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_batch.R
\name{doi_batch}
\alias{doi_batch}
\title{Process a list of dois from a file
and add bibtex file to dir}
\usage{
doi_batch(file, dir)
}
\arguments{
\item{file}{File containing list of doi identifiers}

\item{dir}{Directory in which to save the bibtex citation on success}
}
\value{
Result of doi_to_bibtex for each entry in file
}
\description{
Process a list of dois from a file
and add bibtex file to dir
}
