% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{doi_from_bibtex}
\alias{doi_from_bibtex}
\title{Retrieve a DOI from a bibentry object}
\usage{
doi_from_bibtex(bib)
}
\arguments{
\item{bib}{Bibtex object}
}
\value{
String containing doi
}
\description{
Retrieve a DOI from a bibentry object
}
