% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_lookup.R
\name{doi_lookup}
\alias{doi_lookup}
\title{Look up a DOI and return citation as a bibtex entry}
\usage{
doi_lookup(doi)
}
\arguments{
\item{doi}{DOI identifier (or DOI hyperlink)}
}
\value{
Character vector containing bibtex citation, or NULL if no entry found
}
\description{
This is a wrapper function that first queries crossref
using `doi_lookup_crossref`. If no result is found, it
then tries querying dx.doi.org directly using `doi_lookup_direct`.
}
