% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_lookup.R
\name{doi_lookup_crossref}
\alias{doi_lookup_crossref}
\title{Lookup a doi on crossref.org and return the citation in bibtex format}
\usage{
doi_lookup_crossref(doi)
}
\arguments{
\item{doi}{DOI identifier (or DOI hyperlink)}
}
\value{
Character vector containing bibtex citation, or NULL if no entry found
}
\description{
Lookup a doi on crossref.org and return the citation in bibtex format
}
