% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_parse.R
\name{doi_strip_url}
\alias{doi_strip_url}
\title{Remove the URL part (i.e. "http://dx.doi.org/") from a doi}
\usage{
doi_strip_url(doi)
}
\arguments{
\item{doi}{DOI, with or without URL part}
}
\value{
string with URL part removed
}
\description{
Remove the URL part (i.e. "http://dx.doi.org/") from a doi
}
