% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{doi_to_bibtex}
\alias{doi_to_bibtex}
\title{Retrieve the bibtex citation for a given DOI and save to a file.
Not vectorised!}
\usage{
doi_to_bibtex(doi, dir = "bibtex")
}
\arguments{
\item{doi}{DOI to look up and and save to file}

\item{dir}{Directory in which to save the bibtex file (default: "bibtex")}
}
\value{
The result of the rcrossref::cr_cn lookup
}
\description{
Retrieve the bibtex citation for a given DOI and save to a file.
Not vectorised!
}
