% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_parse.R
\docType{data}
\name{doiRegex}
\alias{doiRegex}
\title{https://github.com/regexhq/doi-regex
http://www.doi.org/doi_handbook/2_Numbering.html#2.2}
\format{An object of class \code{character} of length 1.}
\usage{
doiRegex
}
\description{
https://github.com/regexhq/doi-regex
http://www.doi.org/doi_handbook/2_Numbering.html#2.2
}
\keyword{datasets}
