% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_parse.R
\name{extract_doi}
\alias{extract_doi}
\title{Extract all DOI from a character object}
\usage{
extract_doi(string, as_vector = TRUE)
}
\arguments{
\item{string}{String from which to extract all doi entries}

\item{as_vector}{A single logical value. If TRUE (default), calls `unlist` to return to a single character vector. If FALSE, returns a list of character vectors, one for each element of `string`}
}
\description{
Extract all DOI from a character object
}
