% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_parse.R
\name{is_doi}
\alias{is_doi}
\title{Test whether string contains a doi}
\usage{
is_doi(string)
}
\arguments{
\item{string}{String to test}
}
\value{
Logical indicator TRUE (matches doi specification) or FALSE
}
\description{
Test whether string contains a doi
}
