% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_lookup.R
\name{pdf_extract_doi}
\alias{pdf_extract_doi}
\title{Extract a DOI from the first page of a PDF file}
\usage{
pdf_extract_doi(path)
}
\arguments{
\item{path}{Path specifying a PDF file from which the DOI should be extracted}
}
\value{
Character string with matched DOI, or NA if no DOI found
}
\description{
This function returns the first DOI from the first page
of a given PDF file. It uses `pdftools::pdf_text` to convert the first
page of the PDF to text. The PDF must contain readable text.
}
