% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_lookup.R
\name{pdf_parse_filename}
\alias{pdf_parse_filename}
\title{Parse a filename in format AuthorYYYY-Title.pdf}
\usage{
pdf_parse_filename(fname)
}
\arguments{
\item{fname}{File name in format "AuthorYYYY-Title.pdf" (minor variations allowed)}
}
\value{
tibble with columns `filename`, `author`, `year` and `title`.
}
\description{
Parse a filename in format AuthorYYYY-Title.pdf
}
