% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bib.R
\name{read_bib}
\alias{read_bib}
\title{Read bibliographic data}
\usage{
read_bib(database = NULL, extension = NULL, directory = NULL,
  col.names = FALSE, col.names.style = NULL, fileEncoding = "utf-8")
}
\arguments{
\item{database}{\code{WoS} or \code{Scopus} data base.}

\item{extension}{\code{txt} or \code{csv} extension.}

\item{directory}{the path to find the files, if the files are in a diferrent directory.}

\item{col.names}{logical, if TRUE return optional name of column to the data.}

\item{col.names.style}{\code{uppercase}, \code{lower} and \code{good4read}.}
}
\value{
a data frame with the data imported.
}
\description{
Read data from Web of Science 'WoS' and Scopus. It's possible read one o more
files and set the colunm names.
}
\details{
Write the details about the functions here!
Can be in more that one line.
}
\examples{
\dontrun{Import data from Web of Science (WoS)}
p <- read_bib(
             database='WoS', 
             extension='txt',
             directory='/Users/roney/data/WoS tab-delimited Windows',
             col.names=TRUE,
             col.names.style='good4read'
             )
names(p);dim(p)
\dontrun{or for a scopus csv data}
p <- read_bib( database='Scopus', extension='csv', col.names=TRUE, col.names.style='good4read')
names(p);dim(p)
}
\author{
Roney Fraga Souza, Winicius Sabino
}
\seealso{
\code{read.table}.
}

