% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{read_bib_dir}
\alias{read_bib_dir}
\title{Read a directory of bibtex files and return a list of bibentry objects
Files that cannot be read are moved to a subdirectory "bad"}
\usage{
read_bib_dir(dir)
}
\arguments{
\item{dir}{Directory containing bibtex files}
}
\value{
List of bibtex objects
}
\description{
Read a directory of bibtex files and return a list of bibentry objects
Files that cannot be read are moved to a subdirectory "bad"
}
