% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{rename_bibtex_doi}
\alias{rename_bibtex_doi}
\title{Rename a bibtex file to either the cleaned DOI or,
if not present, the bibkey}
\usage{
rename_bibtex_doi(bibfile)
}
\arguments{
\item{bibfile}{File containing a single bibtex entry}
}
\value{
Nothing
}
\description{
Rename a bibtex file to either the cleaned DOI or,
if not present, the bibkey
}
