% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_search.R
\name{search_bibentries}
\alias{search_bibentries}
\title{Search for pattern in a list of bibtex objects}
\usage{
search_bibentries(bib, pattern, fuzzy = TRUE, ignore.case = TRUE,
  regexp = TRUE, ...)
}
\arguments{
\item{bib}{List of bibtex objects}

\item{pattern}{to search for}

\item{fuzzy}{Whether to use fuzzy matching (agrep) (default: true)}

\item{ignore.case}{Whether to ignore case (default: true)}

\item{regexp}{Whether pattern is a regular expression (default: true)}

\item{...}{Further parameters passed to agrep or grep}
}
\value{
List of bibtex objects matching the pattern
}
\description{
Search for pattern in a list of bibtex objects
}
