% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_search.R
\name{search_library}
\alias{search_library}
\title{Search library (i.e. combination of search_bibtex and search_files)}
\usage{
search_library(pattern, bib, files, fuzzy = TRUE, ignore.case = TRUE,
  regexp = TRUE, ...)
}
\arguments{
\item{pattern}{Pattern to search for}

\item{bib}{List of bibtex objects}

\item{files}{data.frame with file metadata}

\item{fuzzy}{Whether to use fuzzy matching (agrep) (default: true)}

\item{ignore.case}{Whether to ignore case (default: true)}

\item{regexp}{Whether pattern is a regular expression (default: true)}

\item{...}{Further parameters passed to agrep or grep}
}
\value{
rows of data frame matching pattern
}
\description{
Search library (i.e. combination of search_bibtex and search_files)
}
