% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_bib.R
\name{split_bib}
\alias{split_bib}
\title{Split fields and return a data frame}
\usage{
split_bib(x, by = ";", address = NULL)
}
\arguments{
\item{\code{x}}{a vector with text}

\item{\code{by}}{the character used to separate the fields}

\item{\code{address}}{can be 'institution', 'country' or 'whole field'}
}
\value{
a data frame from the divided string
}
\description{
Split fields as Author Address, Authors, Keywords
and return a data frame.
}
\details{
Write the details about the functions here!
Can be in more that one line.
}
\examples{
\dontrun{Split the author address from Web of Science (WoS)}
c1 <- split_bib(p$AuthorAddress, by=';', address='whole field')
# get the country 
country <- split_bib(p$AuthorAddress, by=';', address='country')

# get the institution of research of the authors
inst <- split_bib(p$AuthorAddress, by=';', address='institution')
}
\author{
Roney Fraga Souza, Winicius Sabino
}

