% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{split_bibtex}
\alias{split_bibtex}
\title{Parse a bibtex file and save each entry to a separate file}
\usage{
split_bibtex(bibfile, dir_to = NULL)
}
\arguments{
\item{bibfile}{File containing (possibly) multiple bibtex entries}

\item{dir_to}{Directory in which to save the individual entries (default: NULL, use same directory as source file)}
}
\description{
Parse a bibtex file and save each entry to a separate file
}
